Clean up runtime boundary residue
