+++ src/app/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) use crate::compat_app_contract::*;
+pub(crate) use crate::app_core::native_shell::runtime_contract::*;
 
 pub mod controller;
 pub mod state;
