+++ src/app_core/actions/tests.rs
@@ -10,8 +10,7 @@ use crate::app_core::native_bridge::{
 };
 use crate::gui_test::{GuiAivAssertion, GuiAivStep, gui_aiv_suite_manifest};
 use std::collections::BTreeSet;
-use std::fs;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 #[test]
 fn catalog_contains_every_action_kind_exactly_once() {
@@ -87,30 +86,6 @@ fn representative_actions_round_trip_through_kind_matcher() {
     for kind in GUI_ACTION_CATALOG.iter().map(|entry| entry.kind) {
         let action = representative_action_for_kind(kind);
         assert_eq!(action_kind(&action), kind);
-    }
-}
-
-#[test]
-/// Handles representative actions round trip through compat conversion.
-fn representative_actions_round_trip_through_compat_conversion() {
-    for entry in GUI_ACTION_CATALOG {
-        let action = representative_action_for_kind(entry.kind);
-        let compat_action: crate::app_core::native_shell::runtime_contract::UiAction =
-            action.clone().into();
-        let round_trip = crate::app_core::actions::NativeUiAction::from(compat_action.clone());
-        assert_eq!(
-            round_trip, action,
-            "native -> compat -> native conversion changed {}",
-            entry.action_id
-        );
-
-        let compat_round_trip: crate::app_core::native_shell::runtime_contract::UiAction =
-            round_trip.into();
-        assert_eq!(
-            compat_round_trip, compat_action,
-            "compat conversion is not stable for {}",
-            entry.action_id
-        );
     }
 }
 
@@ -247,664 +222,131 @@ fn native_app_default_keeps_sempal_product_labels_owned_in_app_core() {
 }
 
 #[test]
-fn native_action_exports_are_owned_in_app_core() {
-    let manifest_dir = Path::new(env!("CARGO_MANIFEST_DIR"));
-    let actions_mod =
-        fs::read_to_string(manifest_dir.join("src/app_core/actions/mod.rs")).expect("actions mod");
-
-    assert!(
-        !actions_mod.contains("pub type NativeUiAction = radiant::runtime_contract::legacy_shell"),
-        "NativeUiAction must stay Sempal-owned, with compatibility conversion at the runtime boundary"
-    );
-    assert!(
-        !actions_mod.contains("pub type NativeAppModel = radiant::runtime_contract::legacy_shell"),
-        "NativeAppModel must stay Sempal-owned, with compatibility conversion at the runtime boundary"
-    );
-    assert!(
-        !actions_mod
-            .contains("pub type NativeDirtySegments = radiant::runtime_contract::legacy_shell"),
-        "NativeDirtySegments must stay Sempal-owned, with compatibility conversion at the runtime boundary"
-    );
-    assert!(
-        actions_mod.contains("mod native_shell_actions;")
-            && actions_mod.contains("mod native_shell_bridge;")
-            && actions_mod.contains("pub(crate) mod native_shell_dtos;"),
-        "Sempal-owned action, bridge, and projection DTO modules must remain explicit"
-    );
-    assert!(
-        !actions_mod.contains("pub use radiant::runtime_contract::legacy_shell::NativeAppBridge"),
-        "NativeAppBridge must stay Sempal-owned, with compatibility conversion at the runtime boundary"
-    );
-
-    let local_contract =
-        fs::read_to_string(manifest_dir.join("src/app_core/native_shell/runtime_contract.rs"))
-            .expect("local compatibility contract");
-    assert!(
-        local_contract.contains("Sempal-owned compatibility contract")
-            && local_contract.contains("pub use actions::{BrowserTriageTarget, UiAction};")
-            && local_contract.contains("pub use shell::{"),
-        "Sempal must have a local compatibility contract before the Radiant legacy shell facade can be retired"
-    );
-    assert!(
-        !local_contract.contains("runtime_artifacts")
-            && !local_contract.contains("NativeRunReport")
-            && !local_contract.contains("NativeRuntimeArtifacts"),
-        "runtime artifact envelopes belong in gui_runtime, not the local compatibility DTO contract"
-    );
-
-    let native_dtos =
-        fs::read_to_string(manifest_dir.join("src/app_core/actions/native_shell_dtos.rs"))
-            .expect("native shell dtos");
-    let native_actions =
-        fs::read_to_string(manifest_dir.join("src/app_core/actions/native_shell_actions.rs"))
-            .expect("native shell actions");
-    assert!(
-        !native_actions.contains("radiant::runtime_contract::legacy_shell")
-            && !native_actions.contains("runtime_contract::UiAction")
-            && !native_actions.contains("runtime_contract::BrowserTriageTarget"),
-        "Sempal action payload definitions should not import Radiant legacy-shell compatibility types"
-    );
-    let native_action_conversions = native_shell_runtime_sources(&manifest_dir);
-    let native_actions = format!("{native_actions}\n{native_action_conversions}");
-    assert!(
-        !manifest_dir
-            .join("src/app_core/native_shell/composition/runtime/native_vello.rs")
-            .exists(),
-        "Radiant legacy-shell conversion belongs in gui_runtime, not app_core native-shell composition"
-    );
-    for source_path in rust_sources_under(&manifest_dir.join("src/app_core/native_shell")) {
-        let source = fs::read_to_string(&source_path).expect("native shell source");
-        assert!(
-            !source.contains("radiant::runtime_contract::legacy_shell"),
-            "{} must not import Radiant legacy-shell compatibility contracts; keep that adapter in gui_runtime",
-            source_path.display()
+fn native_runtime_contract_round_trips_representative_actions() {
+    for entry in GUI_ACTION_CATALOG {
+        let action = representative_action_for_kind(entry.kind);
+        let runtime_action: crate::app_core::native_shell::runtime_contract::UiAction =
+            action.clone().into();
+        let product_round_trip =
+            crate::app_core::actions::NativeUiAction::from(runtime_action.clone());
+        assert_eq!(
+            product_round_trip, action,
+            "native -> runtime -> native conversion changed {}",
+            entry.action_id
         );
-    }
-    for source_path in rust_sources_under(&manifest_dir.join("src")) {
-        if is_test_source(&source_path) {
-            continue;
-        }
-        let source = fs::read_to_string(&source_path).expect("Sempal source");
-        assert!(
-            !source.contains("radiant::runtime_contract::legacy_shell")
-                && !source.contains("run_legacy_native_vello_app_with_artifacts"),
-            "{} must not depend on Radiant's legacy-shell contracts or runner",
-            source_path.display()
+
+        let runtime_round_trip: crate::app_core::native_shell::runtime_contract::UiAction =
+            product_round_trip.into();
+        assert_eq!(
+            runtime_round_trip, runtime_action,
+            "runtime conversion is not stable for {}",
+            entry.action_id
         );
-    }
-    assert!(
-        !native_dtos.contains("radiant::runtime_contract::legacy_shell")
-            && !native_dtos.contains("runtime_contract::"),
-        "Sempal projection DTO definitions should not import Radiant legacy-shell compatibility types"
-    );
-    let native_hit_testing = fs::read_to_string(
-        manifest_dir.join("src/app_core/native_shell/composition/state/hit_testing/browser.rs"),
-    )
-    .expect("native browser hit testing");
-    let waveform_shell_hit_testing = fs::read_to_string(
-        manifest_dir.join("src/app_core/native_shell/composition/state/hit_testing/waveform.rs"),
-    )
-    .expect("native waveform hit testing");
-    let waveform_hit_testing = fs::read_to_string(
-        manifest_dir
-            .join("src/app_core/native_shell/composition/state/hit_testing/waveform/toolbar.rs"),
-    )
-    .expect("native waveform toolbar hit testing");
-    let waveform_playhead_trail = fs::read_to_string(
-        manifest_dir
-            .join("src/app_core/native_shell/composition/state/motion_overlay/playhead_trail.rs"),
-    )
-    .expect("native waveform playhead trail");
-    let waveform_segment_trail = fs::read_to_string(
-        manifest_dir.join("src/app_core/native_shell/composition/state/waveform_segments/trail.rs"),
-    )
-    .expect("native waveform segment trail");
-    let waveform_segment_overlay = fs::read_to_string(
-        manifest_dir
-            .join("src/app_core/native_shell/composition/state/waveform_segments/overlay.rs"),
-    )
-    .expect("native waveform segment overlay");
-    let waveform_segment_slices = fs::read_to_string(
-        manifest_dir
-            .join("src/app_core/native_shell/composition/state/waveform_segments/slices.rs"),
-    )
-    .expect("native waveform segment slices");
-    let waveform_segment_scrollbar = fs::read_to_string(
-        manifest_dir
-            .join("src/app_core/native_shell/composition/state/waveform_segments/scrollbar.rs"),
-    )
-    .expect("native waveform segment scrollbar");
-    let waveform_segment_fades = fs::read_to_string(
-        manifest_dir.join("src/app_core/native_shell/composition/state/waveform_segments/fades.rs"),
-    )
-    .expect("native waveform segment fades");
-    let waveform_toolbar_helpers =
-        fs::read_to_string(manifest_dir.join(
-            "src/app_core/native_shell/composition/state/toolbar_helpers/waveform_toolbar.rs",
-        ))
-        .expect("native waveform toolbar helpers");
-    let waveform_header_surface = fs::read_to_string(
-        manifest_dir.join("src/app_core/native_shell/composition/waveform_header_surface.rs"),
-    )
-    .expect("native waveform header surface");
-    let waveform_static_frame = fs::read_to_string(
-        manifest_dir.join("src/app_core/native_shell/composition/state/frame_build/waveform.rs"),
-    )
-    .expect("native waveform static frame builder");
-    let waveform_hover_overlay = fs::read_to_string(
-        manifest_dir
-            .join("src/app_core/native_shell/composition/state/frame_build/overlay/waveform.rs"),
-    )
-    .expect("native waveform hover overlay frame builder");
-    let native_options_actions = fs::read_to_string(
-        manifest_dir.join("src/app_core/native_shell/composition/state/options_panel/actions.rs"),
-    )
-    .expect("native options-panel actions");
-    let native_projection_key = fs::read_to_string(
-        manifest_dir
-            .join("src/app_core/native_bridge/projection_cache/projection_key/non_segment.rs"),
-    )
-    .expect("native projection key");
-    assert!(
-        !native_dtos.contains("pub struct RetainedVec"),
-        "retained snapshot storage should use the Radiant-owned generic primitive"
-    );
-    assert!(
-        native_dtos.contains("pub type RetainedVec<T> = retained::RetainedVec<T>;"),
-        "Sempal native DTOs should alias the generic Radiant retained storage primitive"
-    );
-    assert!(
-        !native_dtos.contains("Self::from(runtime_contract::AppModel::default())")
-            && native_dtos
-                .contains("title: String::from(crate::gui_runtime::DEFAULT_NATIVE_WINDOW_TITLE)")
-            && native_dtos.contains("ColumnModel::new(\"Samples\", 0)"),
-        "Sempal app-model defaults should be app-core owned rather than sourced through Radiant legacy-shell DTOs"
-    );
-    assert!(
-        !native_dtos.contains("pub struct AutomationNodeId"),
-        "automation node IDs should use the Radiant-owned generic primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub struct AutomationBounds"),
-        "automation bounds should use the Radiant-owned generic primitive"
-    );
-    assert!(
-        native_dtos.contains("pub type AutomationNodeId = automation::AutomationNodeId;")
-            && native_dtos.contains("pub type AutomationBounds = automation::AutomationBounds;"),
-        "Sempal native automation DTOs should alias generic Radiant automation primitives"
-    );
-    assert!(
-        native_dtos.contains("pub enum AutomationRole")
-            && native_dtos.contains("WaveformRegion")
-            && native_dtos.contains("MapCanvas")
-            && native_dtos.contains("MapPoint"),
-        "Sempal should own product automation role names at the app boundary"
-    );
-    assert!(
-        native_action_conversions
-            .contains("gui_automation::AutomationRole::TimelineRegion => Self::WaveformRegion")
-            && native_action_conversions
-                .contains("gui_automation::AutomationRole::SpatialCanvas => Self::MapCanvas")
-            && native_action_conversions
-                .contains("gui_automation::AutomationRole::SpatialPoint => Self::MapPoint")
-            && !native_action_conversions.contains("runtime_contract::AutomationRole"),
-        "Sempal automation DTO conversion should map generic Radiant roles onto product role names without routing through legacy compat"
-    );
-    assert!(
-        native_action_conversions
-            .contains("\"browser.tab.items\" => String::from(\"browser.tab.samples\")")
-            && native_action_conversions
-                .contains("\"browser.pill_editor\" => String::from(\"browser.tag_sidebar\")")
-            && native_action_conversions.contains("\"browser.pill_editor.option.\"")
-            && native_action_conversions
-                .contains("format!(\"browser.tag_sidebar.normal_tag.{suffix}\")")
-            && native_action_conversions
-                .contains("metadata.insert(String::from(\"normal_tag_labels\"), value);")
-            && native_action_conversions
-                .contains("metadata.insert(String::from(\"tag_state\"), value);")
-            && native_action_conversions
-                .contains("metadata.insert(String::from(\"tag_id\"), value);")
-            && !native_action_conversions
-                .contains("impl From<GuiAutomationSnapshot> for runtime_contract::"),
-        "Sempal automation DTO conversion should map generic Radiant pill-editor nodes and metadata onto product tag-sidebar names without reverse legacy compat conversion"
-    );
-    assert!(
-        !native_dtos.contains("pub struct FrameBuildResult"),
-        "frame feedback should use the Radiant-owned generic primitive"
-    );
-    assert!(
-        native_dtos.contains("pub type FrameBuildResult = frame::FrameBuildResult;"),
-        "Sempal native frame feedback should alias the generic Radiant frame primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub struct NormalizedRangeModel"),
-        "normalized ranges should use the Radiant-owned generic primitive"
-    );
-    assert!(
-        native_dtos.contains("pub type NormalizedRangeModel = range::NormalizedRange;"),
-        "Sempal native range fields should alias the generic Radiant range primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub struct StatusBarModel"),
-        "status chrome should use the Radiant-owned generic primitive"
-    );
-    assert!(
-        native_dtos.contains("pub type StatusBarModel = chrome::StatusSegments;"),
-        "Sempal status chrome should alias the generic Radiant chrome primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub struct ProgressOverlayModel")
-            && !native_dtos.contains("pub struct DragOverlayModel"),
-        "feedback overlays should use Radiant-owned generic primitives"
-    );
-    assert!(
-        native_dtos.contains("pub type ProgressOverlayModel = feedback::ProgressOverlay;")
-            && native_dtos.contains("pub type DragOverlayModel = feedback::DragOverlay;"),
-        "Sempal feedback overlays should alias generic Radiant feedback primitives"
-    );
-    assert!(
-        !native_dtos.contains("pub struct ColumnModel")
-            && !native_dtos.contains("pub enum FolderRowKind")
-            && !native_dtos.contains("pub struct FolderRowModel")
-            && !native_dtos.contains("pub struct FolderPaneModel")
-            && !native_dtos.contains("pub struct FolderActionsModel"),
-        "generic list column, editable-row kind, editable-tree row/pane, and editable-tree actions should use Radiant-owned primitives"
-    );
-    assert!(
-        native_dtos.contains("pub type ColumnModel = list::ColumnSummary;")
-            && native_dtos.contains("pub type FolderPaneIdModel = panel::SplitPaneSlot;")
-            && native_dtos
-                .contains("pub type FolderPaneModel = panel::SplitPaneTreePanel<FolderRowModel>;")
-            && native_dtos.contains("pub type SourceRowModel = panel::SplitPaneAssignedRow;")
-            && native_dtos.contains("pub type FolderRowKind = list::EditableRowKind;")
-            && native_dtos.contains("pub type FolderRowModel = list::EditableTreeRow;")
-            && native_dtos.contains("pub type FolderActionsModel = list::EditableTreeActions;")
-            && native_dtos.contains("panel::SplitPaneSidebarState"),
-        "Sempal source/sidebar DTOs should alias generic Radiant list primitives"
-    );
-    assert!(
-        native_dtos.contains("pane.select(&self.upper_folder_pane, &self.lower_folder_pane)"),
-        "Sempal source/sidebar DTOs should route split-pane lookup through Radiant"
-    );
-    assert!(
-        native_options_actions.contains("model.options_panel.preference_state()")
-            && native_options_actions.contains("preferences.toggle_enabled(0)")
-            && native_projection_key.contains("model.preference_state()")
-            && native_projection_key.contains("preferences.toggles.hash"),
-        "Sempal options-panel rendering and cache keys should consume Radiant's generic preference state"
-    );
-    assert!(
-        native_dtos.contains("pub enum BrowserRowProcessingState")
-            && native_dtos.contains("pub struct BrowserRowModel")
-            && native_dtos.contains("pub enum PlaybackAgeFilterChip")
-            && native_dtos.contains("pub enum PlaybackAgeBucket")
-            && !native_dtos.contains("pub type BrowserRowModel = list::ContentListRow;")
-            && !native_dtos.contains("pub type PlaybackAgeFilterChip = list::RecencyFilterChip;")
-            && !native_dtos.contains("pub type PlaybackAgeBucket = list::RecencyBucket;")
-            && !native_dtos.contains("pub enum BrowserTagState")
-            && !native_dtos.contains("pub struct BrowserTagPillModel")
-            && !native_dtos.contains("pub struct BrowserTagSidebarModel")
-            && !native_dtos.contains("pub enum MapRenderModeModel")
-            && !native_dtos.contains("pub struct MapPanelModel")
-            && native_dtos.contains("pub struct WaveformSlicePreviewModel")
-            && native_dtos.contains("pub struct WaveformToolStateModel")
-            && native_dtos.contains(
-                "visualization::TimelineMotionState<WaveformSlicePreviewModel, WaveformToolStateModel>"
-            )
-            && !native_dtos.contains(
-                "pub type WaveformSlicePreviewModel = visualization::TimelineMarkerPreview;"
-            )
-            && !native_dtos.contains(
-                "pub type WaveformToolStateModel = visualization::SignalToolState;"
-            )
-            && !native_dtos.contains("pub enum UpdateStatusModel"),
-        "Sempal browser row/playback-age and waveform workflow DTOs should be host-owned while generic pill, map, and update DTOs stay Radiant-owned"
-    );
-    assert!(
-        native_dtos.contains("pub type BrowserTagState = selection::TriState;")
-            && native_dtos
-                .contains("pub type BrowserTagPillModel = badge::SelectablePill<BrowserTagState>;")
-            && native_dtos.contains(
-                "pub type BrowserTagSidebarModel = badge::PillEditorPanel<BrowserTagState>;"
-            )
-            && native_dtos
-                .contains("pub type MapRenderModeModel = visualization::PointRenderMode;")
-            && native_dtos.contains("pub type MapPanelModel = visualization::SpatialPanel;")
-            && native_dtos
-                .contains("pub type WaveformChannelViewModel = visualization::ChannelViewMode;")
-            && native_dtos.contains("pub type MapPointModel = visualization::SpatialPoint;")
-            && native_dtos.contains("pub type UpdateStatusModel = feedback::UpdateStatus;"),
-        "Sempal native DTOs should alias generic Radiant state primitives"
-    );
-    assert!(
-        !native_dtos.contains("pub struct MapPointModel"),
-        "map point geometry should use the Radiant-owned generic spatial point primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub enum WaveformChannelViewModel"),
-        "waveform channel view should use the Radiant-owned generic channel-view primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub enum FolderPaneIdModel"),
-        "split pane identity should use the Radiant-owned generic panel primitive"
-    );
-    assert!(
-        native_dtos.contains("pub enum FocusContextModel")
-            && native_dtos.contains("SampleBrowser")
-            && native_dtos.contains("SourceFolders"),
-        "Sempal should own focus-context names for product surfaces"
-    );
-    assert!(
-        !native_dtos.contains("pub struct SourceRowModel"),
-        "split pane row projection should use the Radiant-owned generic panel primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub struct UpdatePanelModel"),
-        "update panels should use the Radiant-owned generic feedback primitive"
-    );
-    assert!(
-        native_dtos.contains("pub type UpdatePanelModel = feedback::UpdatePanel;"),
-        "Sempal native update panels should alias the generic Radiant feedback primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub struct ConfirmPromptModel"),
-        "confirm prompts should use the Radiant-owned generic prompt primitive with a Sempal prompt kind"
-    );
-    assert!(
-        native_dtos
-            .contains("pub type ConfirmPromptModel = feedback::ConfirmPrompt<ConfirmPromptKind>;"),
-        "Sempal native confirm prompts should alias the generic Radiant prompt primitive"
-    );
-    assert!(
-        native_dtos.contains("pub enum ConfirmPromptKind")
-            && native_dtos.contains("DestructiveEdit")
-            && native_dtos.contains("BrowserRename")
-            && native_dtos.contains("FolderRename")
-            && native_dtos.contains("OptionsDefaultIdentifier"),
-        "Sempal should own product prompt-kind names at the app boundary"
-    );
-    assert!(
-        native_action_conversions.contains(
-            "runtime_contract::ConfirmPromptKind::DestructiveOperation => Self::DestructiveEdit"
-        ) && native_action_conversions
-            .contains("runtime_contract::ConfirmPromptKind::RenameContent => Self::BrowserRename")
-            && native_action_conversions.contains(
-                "runtime_contract::ConfirmPromptKind::RenameNavigationItem => Self::FolderRename"
-            )
-            && native_action_conversions.contains(
-                "runtime_contract::ConfirmPromptKind::CreateNavigationItem => Self::FolderCreate"
-            )
-            && native_action_conversions
-                .contains("ConfirmPromptKind::DestructiveEdit => Self::DestructiveOperation")
-            && native_action_conversions
-                .contains("ConfirmPromptKind::BrowserRename => Self::RenameContent")
-            && native_action_conversions
-                .contains("ConfirmPromptKind::FolderRename => Self::RenameNavigationItem")
-            && native_action_conversions
-                .contains("ConfirmPromptKind::FolderCreate => Self::CreateNavigationItem"),
-        "Sempal prompt DTO conversion should map product prompt names onto generic Radiant intents"
-    );
-    assert!(
-        !native_dtos.contains("pub enum AudioEngineChipStateModel"),
-        "audio chip health should use the Radiant-owned generic health-state primitive"
-    );
-    assert!(
-        native_dtos.contains("pub type AudioEngineChipStateModel = feedback::HealthState;"),
-        "Sempal native audio chip state should alias the generic Radiant health primitive"
-    );
-    assert!(
-        !native_dtos.contains("pub struct AudioOptionItemModel")
-            && !native_dtos.contains("pub struct AudioFieldModel"),
-        "audio picker item and summary field containers should use Radiant-owned generic form primitives"
-    );
-    assert!(
-        native_dtos
-            .contains("pub type AudioOptionItemModel = form::OptionItem<AudioOptionValueModel>;")
-            && native_dtos.contains("pub type AudioFieldModel = form::SummaryField;"),
-        "Sempal native audio picker DTOs should alias generic Radiant form primitives"
-    );
-    assert!(
-        native_action_conversions
-            .contains("impl From<runtime_contract::PairedDevicePanelModel> for AudioEngineModel")
-            && native_action_conversions.contains(
-                "impl From<AudioEngineModel> for runtime_contract::PairedDevicePanelModel"
-            )
-            && native_action_conversions.contains("audio_engine: value.paired_device.into()")
-            && native_action_conversions.contains("paired_device: value.audio_engine.into()")
-            && native_dtos.contains("pub fn paired_device_panel(&self) -> &AudioEngineModel"),
-        "Sempal audio DTO conversion should map to Radiant generic paired-device model names"
-    );
-    assert!(
-        native_dtos.contains("pub fn active_picker(&self) -> Option<form::PairedPickerTarget>")
-            && native_dtos.contains("pub fn options_for(&self, target: form::PairedPickerTarget)")
-            && !native_dtos.contains(
-                "pub fn active_picker(&self) -> Option<runtime_contract::PairedPickerTargetModel>"
-            )
-            && !native_dtos.contains(
-                "pub fn options_for(&self, target: runtime_contract::PairedPickerTargetModel)"
-            ),
-        "Sempal audio DTO helpers should expose the generic Radiant form target rather than the legacy-shell alias"
-    );
-    assert!(
-        native_actions.contains("pub enum BrowserTagTarget")
-            && native_actions.contains("Trash")
-            && native_actions.contains("Keep"),
-        "Sempal should own browser triage target names"
-    );
-    assert!(
-        native_actions.contains("runtime_contract::UiAction::MoveDiscardedItemsToFolder")
-            && native_actions.contains("Self::MoveTrashedSamplesToFolder")
-            && native_actions.contains("UiAction::MoveTrashedSamplesToFolder")
-            && native_actions.contains("Self::MoveDiscardedItemsToFolder"),
-        "Sempal action conversion should map the product trash action onto Radiant's generic discard action"
-    );
-    assert!(
-        native_actions.contains("runtime_contract::UiAction::FocusLoadedContentInList")
-            && native_actions.contains("Self::FocusLoadedSampleInBrowser")
-            && native_actions.contains("UiAction::FocusLoadedSampleInBrowser")
-            && native_actions.contains("Self::FocusLoadedContentInList"),
-        "Sempal action conversion should map the product loaded-sample focus action onto Radiant's generic loaded-content focus action"
-    );
-    assert!(
-        native_actions.contains("runtime_contract::UiAction::SetCompareAnchorFromFocusedContent")
-            && native_actions.contains("Self::SetCompareAnchorFromFocusedBrowserSample")
-            && native_actions.contains("UiAction::SetCompareAnchorFromFocusedBrowserSample")
-            && native_actions.contains("Self::SetCompareAnchorFromFocusedContent"),
-        "Sempal action conversion should map the product compare-anchor action onto Radiant's generic focused-content action"
-    );
-    assert!(
-        native_actions.contains("runtime_contract::UiAction::ToggleContentMark")
-            && native_actions.contains("Self::ToggleBrowserSampleMark")
-            && native_actions.contains("UiAction::ToggleBrowserSampleMark")
-            && native_actions.contains("Self::ToggleContentMark"),
-        "Sempal action conversion should map the product browser mark action onto Radiant's generic content mark action"
-    );
-    assert!(
-        native_actions.contains("runtime_contract::UiAction::ToggleFindSimilarFocusedContent")
-            && native_actions.contains("Self::ToggleFindSimilarFocusedSample")
-            && native_actions.contains("UiAction::ToggleFindSimilarFocusedSample")
-            && native_actions.contains("Self::ToggleFindSimilarFocusedContent"),
-        "Sempal action conversion should map the product find-similar action onto Radiant's generic focused-content action"
-    );
-    assert!(
-        native_actions.contains("runtime_contract::UiAction::NormalizeFocusedContentItem")
-            && native_actions.contains("Self::NormalizeFocusedBrowserSample")
-            && native_actions.contains("UiAction::NormalizeFocusedBrowserSample")
-            && native_actions.contains("Self::NormalizeFocusedContentItem"),
-        "Sempal action conversion should map the product normalize action onto Radiant's generic focused-content normalize action"
-    );
-    assert!(
-        native_actions.contains("runtime_contract::UiAction::PlayRandomContentItem")
-            && native_actions.contains("Self::PlayRandomSample")
-            && native_actions.contains("runtime_contract::UiAction::PlayPreviousRandomContentItem")
-            && native_actions.contains("Self::PlayPreviousRandomSample")
-            && native_actions.contains("UiAction::PlayRandomSample")
-            && native_actions.contains("Self::PlayRandomContentItem")
-            && native_actions.contains("UiAction::PlayPreviousRandomSample")
-            && native_actions.contains("Self::PlayPreviousRandomContentItem"),
-        "Sempal action conversion should map product random-sample actions onto Radiant's generic random-content actions"
-    );
-    assert!(
-        native_actions
-            .contains("runtime_contract::UiAction::FocusSpatialContentItem { content_id }")
-            && native_actions.contains("Self::FocusMapSample")
-            && native_actions.contains("sample_id: content_id")
-            && native_actions.contains("UiAction::FocusMapSample { sample_id }")
-            && native_actions.contains("Self::FocusSpatialContentItem")
-            && native_actions.contains("content_id: sample_id"),
-        "Sempal action conversion should map product map-sample focus onto Radiant's generic spatial-content focus action"
-    );
-    assert!(
-        native_hit_testing.contains("fn map_focus_action(content_id: String) -> UiAction")
-            && native_hit_testing.contains("map_content_action_at_point")
-            && !native_hit_testing.contains("map_sample_action_at_point")
-            && native_hit_testing.contains("UiAction::FocusSpatialContentItem")
-            && native_hit_testing.contains("UiAction::FocusMapSample"),
-        "shared map hit-testing should emit Radiant's generic spatial-content action in the legacy-shell build and Sempal's product action in the app build"
-    );
-    assert!(
-        native_action_conversions
-            .contains("\"focus_spatial_content_item\" => String::from(\"focus_map_sample\")")
-            && !native_action_conversions
-                .contains("\"focus_map_sample\" => String::from(\"focus_spatial_content_item\")"),
-        "Sempal automation DTO conversion should map Radiant's generic spatial-content action id onto the product map-sample action id without reverse legacy compat conversion"
-    );
-    assert!(
-        native_actions.contains("runtime_contract::UiAction::StartContentItemDrag")
-            && native_actions.contains("Self::StartBrowserSampleDrag")
-            && native_actions.contains("runtime_contract::UiAction::UpdateContentItemDrag")
-            && native_actions.contains("Self::UpdateBrowserSampleDrag")
-            && native_actions.contains("runtime_contract::UiAction::FinishContentItemDrag")
-            && native_actions.contains("Self::FinishBrowserSampleDrag")
-            && native_actions.contains("UiAction::StartBrowserSampleDrag")
-            && native_actions.contains("Self::StartContentItemDrag")
-            && native_actions.contains("UiAction::UpdateBrowserSampleDrag")
-            && native_actions.contains("Self::UpdateContentItemDrag")
-            && native_actions.contains("UiAction::FinishBrowserSampleDrag")
-            && native_actions.contains("Self::FinishContentItemDrag"),
-        "Sempal action conversion should map product browser-sample drag actions onto Radiant's generic content-item drag actions"
-    );
-    assert!(
-        native_actions
-            .contains("runtime_contract::UiAction::NormalizeWaveformSelectionOrLoadedContent")
-            && native_actions.contains("Self::NormalizeWaveformSelectionOrSample")
-            && native_actions
-                .contains("runtime_contract::UiAction::CropWaveformSelectionToNewContentItem")
-            && native_actions.contains("Self::CropWaveformSelectionToNewSample")
-            && native_actions.contains("runtime_contract::UiAction::DeleteLoadedWaveformContent")
-            && native_actions.contains("Self::DeleteLoadedWaveformSample")
-            && native_actions.contains("UiAction::NormalizeWaveformSelectionOrSample")
-            && native_actions.contains("Self::NormalizeWaveformSelectionOrLoadedContent")
-            && native_actions.contains("UiAction::CropWaveformSelectionToNewSample")
-            && native_actions.contains("Self::CropWaveformSelectionToNewContentItem")
-            && native_actions.contains("UiAction::DeleteLoadedWaveformSample")
-            && native_actions.contains("Self::DeleteLoadedWaveformContent"),
-        "Sempal action conversion should map product waveform sample actions onto Radiant's generic waveform content actions"
-    );
-    assert!(
-        native_dtos.contains("pub type WaveformSurfaceModel = visualization::TimelineSurfaceState")
-            && native_dtos.contains("pub fn timeline_surface(&self) -> WaveformSurfaceModel"),
-        "Sempal waveform DTOs should expose Radiant's generic timeline surface aggregate"
-    );
-    assert!(
-        native_dtos.contains("visualization::TimelineMotionState<")
-            && native_dtos.contains("pub fn timeline_motion(&self) -> WaveformMotionModel"),
-        "Sempal motion DTOs should expose Radiant's generic timeline motion aggregate"
-    );
-    assert!(
-        waveform_hit_testing.contains("model.signal_chrome()")
-            && waveform_hit_testing.contains("model.signal_tools()")
-            && waveform_hit_testing.contains("model.waveform_presentation()")
-            && waveform_shell_hit_testing.contains("model.waveform.viewport()")
-            && waveform_shell_hit_testing.contains("model.waveform.transport()")
-            && waveform_shell_hit_testing.contains("model.waveform.edit_preview()")
-            && waveform_playhead_trail.contains("model.waveform_transport()")
-            && waveform_playhead_trail.contains("model.waveform_viewport()")
-            && waveform_segment_trail.contains("model.waveform_transport()")
-            && waveform_segment_trail.contains("model.waveform_viewport()")
-            && waveform_segment_overlay.contains("model.waveform_transport()")
-            && waveform_segment_overlay.contains("model.waveform_viewport()")
-            && waveform_segment_overlay.contains("model.waveform_edit_preview()")
-            && waveform_segment_overlay.contains("model.waveform_presentation()")
-            && waveform_segment_overlay.contains("model.waveform_image_preview()")
-            && waveform_segment_slices.contains("model.waveform_viewport()")
-            && waveform_segment_scrollbar.contains("model.waveform_viewport()")
-            && waveform_segment_fades.contains("TimelineCoordinateMapper")
-            && waveform_toolbar_helpers.contains("model.signal_chrome()")
-            && waveform_toolbar_helpers.contains("model.signal_tools()")
-            && waveform_toolbar_helpers.contains("model.waveform_presentation()")
-            && waveform_toolbar_helpers.contains("model.waveform_image_preview()")
-            && waveform_header_surface.contains("model.waveform_transport()")
-            && waveform_header_surface.contains("model.waveform_viewport()")
-            && waveform_header_surface.contains("model.waveform_presentation()")
-            && waveform_header_surface.contains("model.waveform_image_preview()")
-            && waveform_header_surface.contains("model.signal_chrome()")
-            && waveform_static_frame.contains("model.waveform_chrome.signal_tools()")
-            && waveform_static_frame.contains("model.waveform.presentation()")
-            && waveform_static_frame.contains("model.waveform.viewport()")
-            && waveform_static_frame.contains("model.waveform.transport()")
-            && waveform_static_frame.contains("model.waveform.image_preview()")
-            && waveform_hover_overlay.contains("model.waveform_chrome.signal_chrome()")
-            && waveform_hover_overlay.contains("model.waveform_chrome.signal_tools()")
-            && waveform_hover_overlay.contains("model.waveform.presentation()"),
-        "Sempal waveform composition should consume Radiant's generic motion parts on hot paths"
-    );
-    assert!(
-        native_hit_testing.contains("fn focused_similarity_action() -> UiAction")
-            && native_hit_testing.contains("UiAction::ToggleFindSimilarFocusedContent")
-            && native_hit_testing.contains("UiAction::ToggleFindSimilarFocusedSample"),
-        "shared browser hit-testing should emit Radiant's generic focused-content action in the legacy-shell build and Sempal's product action in the app build"
-    );
-    assert!(
-        !native_dtos.contains("pub struct FolderRecoveryModel"),
-        "folder recovery counters should use the Radiant-owned generic recovery summary primitive"
-    );
-    assert!(
-        native_dtos.contains("pub type FolderRecoveryModel = feedback::RecoverySummary;"),
-        "Sempal native folder recovery DTOs should alias the generic Radiant feedback primitive"
-    );
-    assert!(
-        native_dtos.contains("use radiant::gui::invalidation;")
-            && native_dtos.contains("mask: invalidation::RetainedSegmentMask")
-            && native_dtos.contains("mask: invalidation::RetainedSegmentMask::from_bits")
-            && native_dtos.contains("invalidation::RetainedSegmentRevisions"),
-        "Sempal native dirty segments should wrap the generic Radiant retained segment primitives"
-    );
-
-    let radiant_legacy_shell = manifest_dir.join("vendor/radiant/src/compat/legacy_shell");
-    assert!(
-        !radiant_legacy_shell.exists(),
-        "Radiant legacy-shell compatibility contracts should be removed; Sempal must own native action/model compatibility locally"
-    );
-
-    let native_contract_sources = [
-        manifest_dir.join("src/compat_app_contract.rs"),
-        manifest_dir.join("src/app_core/native_shell/composition/runtime/actions/mod.rs"),
-        manifest_dir.join("src/app_core/native_shell/composition/runtime/motion.rs"),
-        manifest_dir.join("src/app_core/native_shell/composition/runtime/shell.rs"),
-    ];
-    let forbidden_native_exports = [
-        "pub type NativeUiAction",
-        "pub enum NativeUiAction",
-        "pub type NativeAppModel",
-        "pub struct NativeAppModel",
-        "pub type NativeDirtySegments",
-        "pub struct NativeDirtySegments",
-        "pub type NativeSourcesPanelModel",
-        "pub struct NativeSourcesPanelModel",
-    ];
-    for source_path in native_contract_sources {
-        let source = fs::read_to_string(&source_path).expect("native contract source");
-        for forbidden in forbidden_native_exports {
-            assert!(
-                !source.contains(forbidden),
-                "{} must not define Sempal-owned {forbidden}",
-                source_path.display()
-            );
-        }
     }
 }
 
+#[test]
+fn native_runtime_contract_preserves_intentional_semantic_translations() {
+    use crate::app_core::actions::{NativeBrowserTagTarget, NativeUiAction};
+    use crate::app_core::native_shell::runtime_contract;
+
+    let action_cases = [
+        (
+            NativeUiAction::FocusLoadedSampleInBrowser,
+            runtime_contract::UiAction::FocusLoadedContentInList,
+        ),
+        (
+            NativeUiAction::SetCompareAnchorFromFocusedBrowserSample,
+            runtime_contract::UiAction::SetCompareAnchorFromFocusedContent,
+        ),
+        (
+            NativeUiAction::ToggleBrowserSampleMark,
+            runtime_contract::UiAction::ToggleContentMark,
+        ),
+        (
+            NativeUiAction::ToggleFindSimilarFocusedSample,
+            runtime_contract::UiAction::ToggleFindSimilarFocusedContent,
+        ),
+        (
+            NativeUiAction::NormalizeFocusedBrowserSample,
+            runtime_contract::UiAction::NormalizeFocusedContentItem,
+        ),
+        (
+            NativeUiAction::PlayRandomSample,
+            runtime_contract::UiAction::PlayRandomContentItem,
+        ),
+        (
+            NativeUiAction::PlayPreviousRandomSample,
+            runtime_contract::UiAction::PlayPreviousRandomContentItem,
+        ),
+        (
+            NativeUiAction::MoveTrashedSamplesToFolder,
+            runtime_contract::UiAction::MoveDiscardedItemsToFolder,
+        ),
+    ];
+
+    for (product_action, runtime_action) in action_cases {
+        let converted_runtime: runtime_contract::UiAction = product_action.clone().into();
+        assert_eq!(converted_runtime, runtime_action);
+        let converted_product = NativeUiAction::from(runtime_action.clone());
+        assert_eq!(converted_product, product_action);
+    }
+
+    assert_eq!(
+        runtime_contract::BrowserTriageTarget::from(NativeBrowserTagTarget::Trash),
+        runtime_contract::BrowserTriageTarget::Negative
+    );
+    assert_eq!(
+        NativeBrowserTagTarget::from(runtime_contract::BrowserTriageTarget::Positive),
+        NativeBrowserTagTarget::Keep
+    );
+}
+
+#[test]
+fn post_cutover_compatibility_facades_are_absent() {
+    let manifest_dir = Path::new(env!("CARGO_MANIFEST_DIR"));
+
+    assert!(
+        !manifest_dir.join("src/compat_app_contract.rs").exists(),
+        "the temporary local runtime-contract facade should be deleted"
+    );
+    assert!(
+        !manifest_dir
+            .join("vendor/radiant/src/compat/legacy_shell")
+            .exists(),
+        "Radiant's removed legacy compatibility tree must stay absent"
+    );
+}
+
+#[test]
+fn native_projection_dtos_keep_product_defaults_and_generic_primitives() {
+    let model = crate::app_core::actions::NativeAppModel::default();
+
+    assert_eq!(model.columns[1].title, "Samples");
+    assert_eq!(model.browser_chrome.samples_tab_label, "Samples");
+    assert_eq!(model.browser_chrome.map_tab_label, "Similarity map");
+    assert_eq!(
+        model.browser_chrome.search_placeholder,
+        "Search samples (Ctrl+F)"
+    );
+    assert_eq!(
+        model
+            .sources
+            .folder_pane(crate::app_core::actions::NativeFolderPaneIdModel::Upper),
+        &model.sources.upper_folder_pane
+    );
+    assert_eq!(
+        model
+            .sources
+            .folder_pane(crate::app_core::actions::NativeFolderPaneIdModel::Lower),
+        &model.sources.lower_folder_pane
+    );
+
+    let preferences = model.options_panel.preference_state();
+    assert_eq!(preferences.toggles.len(), 4);
+    assert_eq!(model.waveform.timeline_surface().markers.len(), 0);
+    assert_eq!(model.waveform_chrome.signal_tools().markers_visible, true);
+}
 fn collect_desktop_aiv_action_ids(
     cases: &[crate::gui_test::GuiAivCase],
     out: &mut BTreeSet<String>,
@@ -934,50 +376,4 @@ fn collect_desktop_aiv_assertion_action_ids(
         };
         out.insert(action_id.clone());
     }
-}
-
-fn rust_sources_under(root: &Path) -> Vec<PathBuf> {
-    let mut sources = Vec::new();
-    collect_rust_sources(root, &mut sources);
-    sources.sort();
-    sources
-}
-
-fn native_shell_runtime_sources(manifest_dir: &Path) -> String {
-    let mut sources =
-        fs::read_to_string(manifest_dir.join("src/gui_runtime/native_shell_runtime.rs"))
-            .expect("native shell runtime facade");
-    for path in [
-        "src/gui_runtime/native_shell_runtime/action_mapping.rs",
-        "src/gui_runtime/native_shell_runtime/automation.rs",
-        "src/gui_runtime/native_shell_runtime/bridge.rs",
-        "src/gui_runtime/native_shell_runtime/input_routing.rs",
-        "src/gui_runtime/native_shell_runtime/launch.rs",
-        "src/gui_runtime/native_shell_runtime/model_mapping.rs",
-    ] {
-        sources.push('\n');
-        sources.push_str(&fs::read_to_string(manifest_dir.join(path)).expect(path));
-    }
-    sources
-}
-
-fn collect_rust_sources(path: &Path, out: &mut Vec<PathBuf>) {
-    let Ok(entries) = fs::read_dir(path) else {
-        return;
-    };
-    for entry in entries.flatten() {
-        let path = entry.path();
-        if path.is_dir() {
-            collect_rust_sources(&path, out);
-        } else if path.extension().and_then(|ext| ext.to_str()) == Some("rs") {
-            out.push(path);
-        }
-    }
-}
-
-fn is_test_source(path: &Path) -> bool {
-    path.file_name().and_then(|name| name.to_str()) == Some("tests.rs")
-        || path
-            .components()
-            .any(|component| component.as_os_str() == "tests")
 }
