+++ src/lib.rs
@@ -45,7 +45,6 @@ mod app_dirs_tests;
 pub mod audio;
 /// Shared helpers used by companion binaries such as the installer and updater helper.
 pub mod companion_apps;
-mod compat_app_contract;
 /// Internal helpers for parsing environment-flag booleans.
 mod env_flags;
 /// Platform helpers for copying files to the clipboard.
